Příprava na M Assembler; logic.php -> util.php
